requirements.txt
flet==0.15.0
manim==0.18.0
numpy==1.24.3
Pillow==9.2.0
